Rename message handlers module
